Moved super → superpauli logic to superop_reps.
